Rename BoardAPi to UserApi
